label some strings in wonder mail
