+++ ld_script.txt
@@ -96,7 +96,6 @@ SECTIONS {
         asm/code_801D760.o(.text);
         src/code_8027C84_pre.o(.text);
         asm/code_8027C84.o(.text);
-        src/code_8028BF0.o(.text);
         src/code_8028DE8.o(.text);
         asm/code_8028DE8.o(.text);
         asm/personality_test.o(.text);
